AirBnB clone v2

+++ 100-clean_web_static.py
@@ -1,10 +1,10 @@
 #!/usr/bin/python3
-""" Deletes out-of-date archives from atatic files """
+""" Deletes out-of-date archives from static files """
 import os
 from fabric.api import *
 
 env.user = 'ubuntu'
-env.key_filename = "~/.ssh/id_rsa"
+env.key_filename = "~/id_rsa"
 env.hosts = ['34.232.53.167', '54.89.195.92']
 
 
@@ -24,14 +24,16 @@ def do_clean(number=0):
         print("The 'versions' directory does not exist.")
         return False
 
-    [archives.pop() for _ in range(number)]
-    with lcd("versions"):
-        [local("rm ./{}".format(a)) for a in archives]
+    if len(archives) > number:
+        [archives.pop(0) for _ in range(len(archives) - number)]
+        with lcd("versions"):
+            [local("rm ./{}".format(a)) for a in archives]
 
     with cd("/data/web_static/releases"):
         archives = run("ls -tr").split()
         archives = [a for a in archives if "web_static_" in a]
-        [archives.pop() for _ in range(number)]
-        [run("rm -rf ./{}".format(a)) for a in archives]
+        if len(archives) > number:
+            [archives.pop(0) for _ in range(len(archives) - number)]
+            [run("rm -rf ./{}".format(a)) for a in archives]
 
     return True
